Initial demo server
Using json_rpc_2 attempt to implement the wire protocol for the
language server through stdin/stdout.

Implement StdIOStreamChange - a StreamChannel that reads from stdin and
writes to stdout, does byte<->String conversion, and manages the headers
defined by the language server protocol. The implementation is
optimistic about the correctness of input, for instance that it has a
header named exactly 'Content-Length'.

An extremely basic proof of concent server using JsonRpc and the wire
protocol, without implementing any of the langauge server apis.