Don't try to set lineMode on stdin
On 1.22 and later this results in an error since stdin is a pipe rather
than a terminal:

StdinException: Error setting terminal line mode, OS Error: Inappropriate ioctl for device, errno = 25

Remove handling for EOT (ascii 4). This was useful for manually using
the parser from the terminal but will not show up for a pipe.