Add uncaught exception handler
Fixes #6

Previously an uncaught async exception would cause the VM to exit and
spit out some garbage ignored by the client because it isn't using the
json rpc protocol.

Now we overwrite a file with the error and keep limping along. Not the
best solution but good enough to make debugging tractable.