Add rootUri as an analysisRoot if it's a package
When opening a .dart file with `cwd` as a large directory the analyzer
freezes up while it churns through all the files. Add some heuristics
for which directories look like a package and only add the `rootUri` if
it matches.

Also searchup for what looks like a package root for `didOpen` to make
it more likely to get full sets of references when jumping to files in
directories outside of the `rootUri`.