Serialize actions per file
When multiple requests were sent rapidly for the same files it was
possible for the content overlays to get out of sync. Use `package:pool`
to ensure that requests for the same file are handled serially. This is
critical for the content overlays, but also should help for other
requests to be sure that they are made for with the most up to day view
of the file.