Stop adding `rootUri` as an analysis root
This has gone back and forth a bit. At the time this was added the
handling of the `didOpen` directories wasn't as nice - we didn't search
up for what looked like the package root. Now that we are doing that it
isn't as necessary to also set the `rootUri` as a root.