Catch exceptions in `applyChange`
There are exceptions when the changes comes in as an append past the end
of the file. We could fix this issue, but we'd have to work around the
same issue in the analysis server by fudging the append to look like
it's an insert before the last newline. For now just catch and log the
exception.