Remove comments
The comment mostly has temporal value which we can find with the `git blame`. It would be OK having a test with this comment to prevent regression, but since the likelihood of removing this is low and the barrier to entry for adding a test is high (since there are so few tests today) I'd prefer to just trust we won't tamper with it for now.