Expose lower level protocol transform
Adds a `lspChannel` top level method which can wrap a
`Stream<List<int>>` and `StreamSink<List<int>>` into a
`StreamChannel<String>` which handles the headers and content length
counting of the LSP protocol, surfacing a stream channel that is the
JSON rpc protocol more directly.

Drop `StdIOStreamChannel`, pass in the `stdin` and `stdout` arguments
instead of hardcoding that into the header handling.